+++ solutions/saas-microservices/README.md
@@ -17,7 +17,7 @@ dashboard.
 
 There are 3 separate applications in this example:
 
-- `dashboard` - A [Next.js](https://nextjs.org/) application to show the UI.
+- `dashboard` - A [Next.js](https://nextjs.org/) application to show the UI. This application also controls the `microfrontnds.json` configuration to route API paths to the other microservices (see below).
 - `api-dashboard` - A [Nitro](https://nitro.build/) backend serving data displayed in the dashboard.
 - `api-users` - A [Hono](https://hono.dev/) backend authenticating users and returning user information.
 
@@ -26,3 +26,22 @@ These all run under the same domain. Paths to each application are routed using 
 - `/api/dashboard/*` - Routes to the `api-dashboard` backend
 - `/api/users/*` - Routes to the `api-users` backend
 - Everything else - Routes to the `dashboard` application
+
+## Running Locally
+
+To run all applications together, run:
+
+```sh
+pnpm dev
+```
+
+A [local development proxy](https://vercel.com/docs/microfrontends/local-development) is automatically run to stitch requests from each application to the local instance of each service.
+
+A single or subset of applications can also be run:
+
+```sh
+pnpm dev:dashboard
+pnpm dev:api-dashboard
+pnpm dev:api-users
+pnpm turbo run dev -F api-users -F api-dashboard
+```
