+++ solutions/saas-microservices/apps/dashboard/src/app/components/recent-activity.tsx
@@ -6,6 +6,7 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
+import { fetchApi } from "@/lib/fetch-api";
 
 interface Activity {
   id: string;
@@ -14,23 +15,14 @@ interface Activity {
   timestamp: string;
 }
 
+interface ActivityResponse {
+  activities: Activity[];
+}
+
 export async function RecentActivity() {
-  const reqHeaders = await headers();
-  const host = reqHeaders.get("host");
-  const isLocalhost = !host || host.includes("localhost");
-  const cookieStore = await cookies();
-  const response = await fetch(
-    `${isLocalhost ? "http://localhost:3024" : "https://saas-microservices-dashboard.vercel.app"}/api/dashboard/activity`,
-    {
-      headers: {
-        Cookie: `saas_microservices_authed_user=${cookieStore.get("saas_microservices_authed_user")?.value}`,
-      },
-    },
-  );
-  const activities = (await response.json()).activities.slice(
-    0,
-    6,
-  ) as Activity[];
+  const activities = (
+    await fetchApi<ActivityResponse>("/api/dashboard/activity")
+  ).activities.slice(0, 6);
 
   return (
     <Card>
